requirements.txt
requests
bs4
dnspython3
python-nmap
cyclonedx-bom
